requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

Pillow==9.3.0

azure-core==1.27.0

azure-functions==1.15.0

azure-storage-blob

azure-servicebus
